# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "CambiarEstadoCajeroRequest": {
        "properties": {
          "accion": {
            "enum": [
              "SUSPENDER",
              "REACTIVAR",
              "RETIRAR"
            ],
            "type": "string"
          }
        },
        "required": [
          "accion"
        ],
        "type": "object"
      },
...
      "InvitacionResponse": {
        "properties": {
          "membresiaId": {
            "format": "uuid",
            "type": "string"
          },
          "pinTemporal": {
            "description": "Se muestra una sola vez. null si la persona ya tenía PIN propio.",
            "example": "482915",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "membresiaId"
        ],
        "type": "object"
      },
      "InvitarCajeroRequest": {
        "properties": {
          "apellidos": {
            "example": "Jacanamejoy",
            "type": "string"
          },
          "celular": {
            "example": "312 456 7890",
            "type": "string"
          },
          "nombres": {
            "example": "Ana Lucía",
            "type": "string"
          },
          "numeroDocumento": {
            "example": "1124500777",
            "type": "string"
          },
          "tipoDocumento": {
            "description": "Código del tipo de documento (catálogo)",
            "example": "CC",
            "type": "string"
          }
        },
        "required": [
          "celular",
          "nombres",
          "tipoDocumento",
          "numeroDocumento"
        ],
        "type": "object"
      },
      "MiembroResponse": {
        "properties": {
          "celular": {
            "example": "+573100000102",
            "nullable": true,
            "type": "string"
          },
          "estado": {
            "enum": [
              "INVITED",
              "ACTIVE",
              "SUSPENDED",
              "REMOVED"
            ],
            "type": "string"
          },
          "membresiaId": {
            "format": "uuid",
            "type": "string"
          },
          "nombre": {
            "example": "Jhon Mutumbajoy",
            "type": "string"
          },
          "roles": {
            "example": [
              "CASHIER"
            ],
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "usuarioId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "membresiaId",
          "usuarioId",
          "nombre",
          "estado",
          "roles"
        ],
        "type": "object"
      },
...
      "PinTemporalResponse": {
        "properties": {
          "pinTemporal": {
            "description": "Se muestra una sola vez",
            "example": "482915",
            "type": "string"
          }
        },
        "required": [
          "pinTemporal"
        ],
        "type": "object"
      },
...
      "RespuestaErrorDto": {
        "properties": {
          "codigo": {
            "example": "HORARIO_SE_CRUZA",
            "type": "string"
          },
          "message": {
            "example": "Ese horario se cruza con otro del mismo día en esta sede.",
            "type": "string"
          },
          "statusCode": {
            "example": 422,
            "type": "number"
          }
        },
        "required": [
          "statusCode",
          "codigo",
          "message"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "El vecino aliado de los negocios del Putumayo. Contrato para la app y el panel.",
    "title": "API de VECI",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/equipo": {
      "get": {
        "operationId": "listarEquipo",
        "parameters": [
          {
            "description": "Id del negocio activo",
            "in": "header",
            "name": "x-veci-comercio",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MiembroResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          },
          "400": {
            "description": "Falta el negocio activo"
          },
          "401": {
            "description": "Sin sesión o sesión cerrada"
          },
          "403": {
            "description": "Requiere el permiso tenancy.manage_staff\n\nEl usuario no trabaja en ese negocio o su rol no alcanza"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Propietarios y cajeros del negocio",
        "tags": [
          "Equipo del negocio"
        ]
      }
    },
    "/equipo/{membresiaId}/estado": {
      "patch": {
        "operationId": "cambiarEstadoCajero",
        "parameters": [
          {
            "description": "Id del negocio activo",
            "in": "header",
            "name": "x-veci-comercio",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "membresiaId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambiarEstadoCajeroRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiembroResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "description": "Falta el negocio activo"
          },
          "401": {
            "description": "Sin sesión o sesión cerrada"
          },
          "403": {
            "description": "Requiere el permiso tenancy.manage_staff\n\nEl usuario no trabaja en ese negocio o su rol no alcanza"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaErrorDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaErrorDto"
                }
              }
            },
            "description": "Cambio no permitido"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Suspende, reactiva o retira",
        "tags": [
          "Equipo del negocio"
        ]
      }
    },
    "/equipo/{membresiaId}/restablecer-pin": {
      "post": {
        "operationId": "restablecerPinCajero",
        "parameters": [
          {
            "description": "Id del negocio activo",
            "in": "header",
            "name": "x-veci-comercio",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "membresiaId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PinTemporalResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "description": "Falta el negocio activo"
          },
          "401": {
            "description": "Sin sesión o sesión cerrada"
          },
          "403": {
            "description": "Requiere el permiso tenancy.reset_staff_pin"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaErrorDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "PIN temporal para un cajero (HU-02-05)",
        "tags": [
          "Equipo del negocio"
        ]
      }
    },
    "/equipo/cajeros": {
      "post": {
        "operationId": "invitarCajero",
        "parameters": [
          {
            "description": "Id del negocio activo",
            "in": "header",
            "name": "x-veci-comercio",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitarCajeroRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitacionResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "description": "Falta el negocio activo"
          },
          "401": {
            "description": "Sin sesión o sesión cerrada"
          },
          "403": {
            "description": "Requiere el permiso tenancy.manage_staff\n\nEl usuario no trabaja en ese negocio o su rol no alcanza"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaErrorDto"
                }
              }
            },
            "description": "Sin cupo, ya está o otro celular"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaErrorDto"
                }
              }
            },
            "description": "Datos inválidos"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invita un cajero por su celular (HU-02-04)",
        "tags": [
          "Equipo del negocio"
        ]
      }
    },
...
  },
  "servers": [],
...
}